Use native prompt token width in test

Copilot-Session: e95773ce-13b6-4f15-b288-ed69843a385c

+++ sdk_v2/cpp/test/internal_api/chat/chat_session_test.cc
@@ -529,7 +529,7 @@ TEST_F(ChatSessionTest, RunMultiTurn) {
 TEST_F(ChatSessionTest, ChatTemplateKwargsControlCachedGeneratorReuse) {
   struct TurnResult {
     std::string text;
-    int prompt_tokens;
+    int64_t prompt_tokens;
   };
 
   auto run_turn = [&](ChatSession& session,
